initial population of docs template with Control Tower content
change the structure of the subnav and add 4 more categories to the using section in the docs-content

reorganise the 4 subcategories for the using section, strongly inspired by our readme but with some tweaks to make sense of the user flow, with a focus on the deploy section for this commit

reorganise the operate section, such as adding more headers to explain the different parts of the user flow. In progress

add a collapsible with arrow for the Using category with the 4 subcategories covering the user journey

add a section in the deploy category to talk about what happens when deploying, in the shape of 2 tables for each IaaS

move the content about getting deployment info to the top of the page

shorten the headers and titles

restructure of the using section in progress, add 2 subsections to the operate category, now need to edit those as well as fixing the invisible deploy category, now disappeared because restructure in progress

structure menu correctly

Put deploy at the top of the page

move credentials config to install section

fix broken link for using ctwr top level menu

add an index file for operation section and content at top of file

Make IaaS provider support more clearly explained

Define a clear IaaS User Permission section

Add more categories to the deploy page, make them match the mind map and add an anchor tag to make them clickable at the top of the page

finalise this deploy page by adding a custom metadata and namespace section, anchored as well

add a section about data protection while in transit or at rest, add a section about security and networking to talk about allow ips and cidr range, add 2 more sections at the bottom of the page to talk about the credentials management, last section about authentication in progress

remove the github auth section as it should be movedto part 2 of the operation category, to match the mind map exactly. Also reformulate the IP range section with allow-ips flag

add one more sentence to explain what the CIDR block is

add a section about the concourse workers persistent public IP. The security and networking section should now mirror what is in our mindmap

add 2 full stops that were missing

add a section at the top about github auth config

add the resource cost monitoring section at the bottom, with a subsection about cost efficiency improvement and another subsection about estimated cost

add a section at the top about worker config

add a section about scaling web nodes vertically

rearrange the configure worker nodes section

finalise the structure on VM config

add a section about patch and global updates

add a category about nats certs rotation, add 2 warning section about downtime caused by upgrades

add a sentence under the improve cost efficiency category

add a few tweaks to the deploy section, notably replacing cup to ctwr when needed, but also ensuring consistency within all flags

replace concourse-up commands to ctwr and making the section about data protection in transit more visually pleasing with better bulleted list

just change the command to run credhub at bottom of page to be consistent with the rest of the commands

resize the subtitles and restructure the nats certs rotation section

mostly add the last section at bottom, about how to get metrics and info

move the github auth to deploy section

rename the deployment file and change the subnav

make deployement section titles more helpful

rename operations section and in the subnav menu

make security networking titles more useful

created scaling section with rename in subnav

restructure operations content

add sections to operations to cover upgrades and monitoring

move the NATs certs section from Scaling file to Monitoring file

push changes about monitoring section:removing the nats certs from the scaling section

add a new file to cover iaas costs in operate section, and add it in the subnav

rename the title of the iaas cost efficiency in the subnave to match the file itself

add detail to installation section about being bale to rename the binary and use of the command as a consequence. Make bottom bulleted list better

Make warning more succinct

Tighten up Requirements and Limitations

Expand on privilige requirements

Explain docs structure

replace to control-tower

replace ctwr to control-tower in the deployment section

replace ctwr to control-tower in the security and networking section

replace ctwr for control-tower in deployment scaling section

replace ctwr to control-tower for the IaaS cost efficiency section

replace ctwr to control-tower for the upgrades section

replace ctwr to control-tower for the monitoring section

replace the ctwr command for control-tower in the destroy section

change the format of commands for consistency so that bnow they appear like in a bash shell

change the release date in the product snapshot of the intro/landing page to March

add iaas flag + change the example commands in install section

rearrange a few things in the deployment section, such as the commands, but also add an image for AWS IAM account permissions table

change the commands to make them look like they are in a bash shell for consistency

change the commands to make them look like they are in a bash shell for consistency

add an example for the db size scaling for consistency

make commands look consistent from the rest of the documentation

change the look of the commands in the section about renaming the binary

--iaas is not optional

make commands look consistent and make them look like they are in a shell

show how to use the environment variable instead of a flag

add missing title to instance type reference table for document consistency

create release notes

add gitignore